+++ 2017/alis/lib/Language.pm
@@ -32,18 +32,6 @@ our @EXPORT_OK = qw(%language check_language);
 # ------------------------------------------------------------------------------
 
 
-sub check_language {
-    # Check that the language pack has been added
-    my @language_list = ("en", "fr", "es", "meme");
-    my $language_selected = $_[0];
-    my $pass = 0;
-    foreach my $lang (@language_list) {
-        if ($lang eq $language_selected) { $pass = 1; }
-    }
-    if ($pass == 1) { return 1; } else { return 0; }
-    1;
-}
-
 
 # Language dictionary
 our %language = (
@@ -55,7 +43,10 @@ our %language = (
         cancel_button => qq{Cancel},
         yes_button    => qq{Yes},
         no_button     => qq{No},
-        # TODO maybe add a back button
+        back_button   => qq{Back},
+
+        wip_title => qq{Work in Progress},
+        wip_message => qq{This section is a work in progress},
 
         welcome_message_title => qq{ALIS - Arch Linux Installation Script},
         welcome_message       => qq{Welcome to ALIS - Arch Linux Installation Script.}. qq{\n}.
@@ -217,6 +208,16 @@ our %language = (
     },
 
     );
+
+
+sub check_language {
+    # Check that the language pack has been added
+    my $language_selected = $_[0];
+    if (defined $language{$language_selected}) {
+        return 1;
+    } else { return 0; }
+}
+
 
 1;